error handling: - UDP syslog reporter constructor cannot fail!   - if sendto fails at runtime, report to console (Printf.eprintf) - TCP syslog reporter constructor can fail if connect fails   - if send fails, log to console, and retry connecting   - this could be more fine grained to only retry on connection refused/no route to host (which may be temporary failures)
The idea is that if connect succeeded once (during construction), there is the
possibility to log messages.  It is best effort, if the destination host is
temporarily down, messages can get lost (although TCP), but will be logged to
console.  In the UDP case, messages can get lost in any way.